+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         <activity android:name=".TempActivity" />
         <activity android:name=".BaseActivity" />
         <activity
-            android:name=".DashboardActivity"
+            android:name=".Dashboard.DashboardActivity"
             android:screenOrientation="landscape" />
         <activity
             android:name=".RestaurantDineIn"
