+++ app/src/main/java/com/example/restaurantdine_in/Dashboard/DashboardFragment.java
@@ -1,18 +1,17 @@
-package com.example.restaurantdine_in;
+package com.example.restaurantdine_in.Dashboard;
 
-import android.app.ActionBar;
-import android.app.TaskStackBuilder;
-import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
-import android.view.Menu;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.Toast;
 
 import androidx.fragment.app.Fragment;
+
+import com.example.restaurantdine_in.Constants;
+import com.example.restaurantdine_in.MenuActivity;
+import com.example.restaurantdine_in.R;
 
 import java.util.ArrayList;
