+++ app/src/main/java/com/example/restaurantdine_in/Dashboard/IDashboardActivityListener.java
@@ -1,4 +1,4 @@
-package com.example.restaurantdine_in;
+package com.example.restaurantdine_in.Dashboard;
 
 public interface IDashboardActivityListener {
 }
